Rename and simplify `actionable-pr-view-file`
