+++ source/refined-github.ts
@@ -96,7 +96,7 @@ import './features/preview-hidden-comments.js';
 import './features/fit-textareas.js';
 import './features/collapsible-content-button.js';
 import './features/resolve-conflicts.js';
-import './features/enable-file-links-in-compare-view.js'; // Must be before more-file-links
+import './features/actionable-pr-view-file.js'; // Must be before more-file-links
 import './features/more-file-links.js';
 import './features/pr-filters.js';
 import './features/quick-file-edit.js';
